chore(tests): isolate parser result regressions
