+++ parser_result_test/parser_result_haskell_quasiquote_test.go
@@ -1,4 +1,4 @@
-package gotreesitter_test
+package parserresult_test
 
 import (
 	"testing"
