+++ parser_result_test/parser_result_rust_contextual_default_test.go
@@ -1,4 +1,4 @@
-package gotreesitter_test
+package parserresult_test
 
 import "testing"
