+++ parser_result_test/parser_result_scala_comment_attachment_test.go
@@ -1,4 +1,4 @@
-package gotreesitter_test
+package parserresult_test
 
 import (
 	"testing"
