+++ parser_result_test/parser_result_trailing_newline_test.go
@@ -1,68 +1,6 @@
-package gotreesitter_test
+package parserresult_test
 
-import (
-	"testing"
-
-	"github.com/odvcencio/gotreesitter"
-	"github.com/odvcencio/gotreesitter/grammars"
-)
-
-func parseByLanguageName(t *testing.T, name, src string) (*gotreesitter.Tree, *gotreesitter.Language) {
-	t.Helper()
-
-	var entry grammars.LangEntry
-	found := false
-	for _, e := range grammars.AllLanguages() {
-		if e.Name == name {
-			entry = e
-			found = true
-			break
-		}
-	}
-	if !found {
-		t.Fatalf("missing language entry %q", name)
-	}
-
-	var backend grammars.ParseBackend
-	found = false
-	for _, report := range grammars.AuditParseSupport() {
-		if report.Name == name {
-			backend = report.Backend
-			found = true
-			break
-		}
-	}
-	if !found {
-		t.Fatalf("missing parse support report %q", name)
-	}
-
-	lang := entry.Language()
-	parser := gotreesitter.NewParser(lang)
-	srcBytes := []byte(src)
-
-	var (
-		tree *gotreesitter.Tree
-		err  error
-	)
-	switch backend {
-	case grammars.ParseBackendTokenSource:
-		if entry.TokenSourceFactory == nil {
-			t.Fatalf("%s: token source backend without factory", name)
-		}
-		tree, err = parser.ParseWithTokenSource(srcBytes, entry.TokenSourceFactory(srcBytes, lang))
-	case grammars.ParseBackendDFA, grammars.ParseBackendDFAPartial:
-		tree, err = parser.Parse(srcBytes)
-	default:
-		t.Fatalf("%s: unsupported backend %q", name, backend)
-	}
-	if err != nil {
-		t.Fatalf("%s parse failed: %v", name, err)
-	}
-	if tree == nil || tree.RootNode() == nil {
-		t.Fatalf("%s parse returned nil tree/root", name)
-	}
-	return tree, lang
-}
+import "testing"
 
 func TestPugTopLevelTagCarriesTrailingNewlineSpan(t *testing.T) {
 	const src = "p hello\n"
